+++ changelog.md
@@ -1,5 +1,5 @@
 # Changelog
 
-**./NameSorter.Tests/SampleDataTest.cs**
-* New file - Runs unit test based on the sample data.
-* Names are hard-coded.
+**./NameSorter/Tasks/ReadInputFile.cs - CheckBytes**
+* Declared 'sizeErrMsg' variable.
+* Error messages are now written using 'ErrorMessages.WriteInputSizeMessage'